D bnormalization

+++ MvcStore/Models/Item.cs
@@ -1,7 +1,5 @@
-using MvcStore.Repo;
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System;
+
 namespace MvcStore.Models
 {
     public class Item
@@ -9,7 +7,7 @@ namespace MvcStore.Models
         public int Id {get; set;}
         public string Name {get; set;}
         public string Description {get; set;}
-        public float Price {get; set;}
+        public double Price {get; set;}
         public int QuantitySold {get; set;}
         [NotMapped]
       public string ImageURL {
